handler_torchserve/deploy-hugging-face-model-to-gcp-main/requirements.txt
python-dotenv
google-cloud-storage
google-cloud
google-cloud-aiplatform
torch-model-archiver
git+https://github.com/huggingface/transformers.git
accelerate==0.21.0
datasets==2.16.1
pandas
typing-extensions==4.3.0
shapely
scikit-image
imgaug
pyclipper
lmdb
tqdm
numpy
visualdl
rapidfuzz
opencv-python==4.6.0.66
opencv-contrib-python==4.6.0.66
cython
lxml
premailer
openpyxl
attrdict
Polygon3
lanms-neo==1.0.2
PyMuPDF
PyYAML
fpdf==1.7.2
pdfkit==1.0.0
wkhtmltopdf==0.2
extract-msg==0.39.0
msgpack==1.0.4
pytesseract==0.3.9
seqeval==1.2.2
pdf2image==1.16.0
lxml==4.9.1
opencv-python-headless==4.6.0.66
google-cloud-vision==3.4.4
torch
torchvision
tabulate
bs4
imgkit
